+++ webknossos/webknossos/dataset/__init__.py
@@ -14,4 +14,5 @@ from .dataset import Dataset
 from .layer import Layer, SegmentationLayer
 from .layer_categories import COLOR_CATEGORY, SEGMENTATION_CATEGORY, LayerCategoryType
 from .mag_view import MagView
+from .sampling_modes import SamplingModes
 from .view import View