+++ webknossos/webknossos/dataset/layer.py
@@ -11,11 +11,11 @@ from typing import TYPE_CHECKING, Dict, List, Optional, Tuple, Union
 import numpy as np
 from upath import UPath
 
+from webknossos.dataset.sampling_modes import SamplingModes
 from webknossos.geometry import BoundingBox, Mag, Vec3Int, Vec3IntLike
 
 from ._array import ArrayException, BaseArray, DataFormat
-from .downsampling_utils import (
-    SamplingModes,
+from ._downsampling_utils import (
     calculate_default_max_mag,
     calculate_mags_to_downsample,
     calculate_mags_to_upsample,
@@ -23,6 +23,7 @@ from .downsampling_utils import (
     downsample_cube_job,
     parse_interpolation_mode,
 )
+from ._upsampling_utils import upsample_cube_job
 from .layer_categories import COLOR_CATEGORY, SEGMENTATION_CATEGORY, LayerCategoryType
 from .properties import (
     LayerProperties,
@@ -32,7 +33,6 @@ from .properties import (
     _properties_floating_type_to_python_type,
     _python_floating_type_to_properties_type,
 )
-from .upsampling_utils import upsample_cube_job
 
 if TYPE_CHECKING:
     from .dataset import Dataset
@@ -637,7 +637,7 @@ class Layer:
 
         Example:
         ```python
-        from webknossos.dataset.downsampling_utils import SamplingModes
+        from webknossos import SamplingModes
 
         # ...
         # let 'layer' be a `Layer` with only `Mag(1)`