+++ webknossos/webknossos/dataset/sampling_modes.py
@@ -1,0 +1,4 @@
+class SamplingModes:
+    ANISOTROPIC = "anisotropic"
+    ISOTROPIC = "isotropic"
+    CONSTANT_Z = "constant_z"