+++ webknossos/Changelog.md
@@ -19,7 +19,7 @@ For upgrade instructions, please check the respective *Breaking Changes* section
 ### Changed
 
 ### Fixed
-
+- Correctly maintain default_view_configuration property when downloading a dataset. [#677](https://github.com/scalableminds/webknossos-libs/pull/677)
 
 ## [0.9.20](https://github.com/scalableminds/webknossos-libs/releases/tag/v0.9.20) - 2022-04-20
 [Commits](https://github.com/scalableminds/webknossos-libs/compare/v0.9.19...v0.9.20)